Move tests related to state management/storage into local-state package
